rename files and folders, use new redux state structure
