+++ lib/main.dart
@@ -4,18 +4,18 @@ import 'package:flutter_redux_navigation/flutter_redux_navigation.dart';
 import 'package:global_configuration/global_configuration.dart';
 import 'package:redux/redux.dart';
 import 'package:redux_thunk/redux_thunk.dart';
-import 'package:scibble/Page/home.dart';
-import 'package:scibble/Page/login.dart';
-import 'package:scibble/Page/onlineView.dart';
-import 'package:scibble/Store/AppState.dart';
-import 'package:scibble/Store/Reducer.dart';
+
+import 'package:scibble/pages/auth/login.dart';
+import 'package:scibble/pages/auth/online_web.dart';
+import 'package:scibble/pages/home.dart';
+import 'package:scibble/redux/store.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+  // Necessary to wait for config retrieval
   await GlobalConfiguration().loadFromAsset('app_settings');
   final _clientId = GlobalConfiguration().get('client_id');
-  final _initalState = AppState(_clientId);
-  // Necessary to wait for config retrieval
+  final _initalState = AppState.initialState(_clientId);
   final Store<AppState> _store = Store<AppState>(
     reducer,
     initialState: _initalState,
@@ -41,7 +41,7 @@ class Scibble extends StatelessWidget {
         navigatorKey: NavigatorHolder.navigatorKey,
         routes: {
           '/': (context) => Login(),
-          '/login/online': (context) => OnlineView(),
+          '/login/online': (context) => OnlineWeb(),
           '/home': (context) => Home(),
         },
       ),
