+++ lib/pages/home.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:redux/redux.dart';
-
-import 'package:scibble/Store/AppState.dart';
+import 'package:scibble/redux/store.dart';
 
 class Home extends StatelessWidget {
   @override
@@ -14,8 +13,9 @@ class Home extends StatelessWidget {
           print('building home');
           return Column(
             children: [
-              Text('Last name: ${store.state.user.lastName}'),
-              Text('Acess token: ${store.state.token.accessToken}'),
+              Text('Last name: ${store.state.auth.userState.user.lastName}'),
+              Text(
+                  'Acess token: ${store.state.auth.tokenState.token.accessToken}'),
             ],
           );
         },
